requirements.txt
firebase-admin>=6.2.0
openai>=1.12.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
logging>=0.5.1.2 